# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Ticket Booking System",
        "title": "Ticket Booking System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Расширенный список мероприятий с фильтрацией по статусу и организатору. / [EN] Extended events list with status and organizer filtering.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-events"
                ],
                "summary": "[RU] Список всех мероприятий (админ) / [EN] All events list (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "[RU] Страница / [EN] Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "[RU] Лимит / [EN] Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "[RU] Статус / [EN] Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "[RU] ID организатора / [EN] Organizer ID",
                        "name": "organizer_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/events/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Общая статистика по всем мероприятиям в системе. / [EN] General statistics for all events in the system.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-events"
                ],
                "summary": "[RU] Статистика мероприятий / [EN] Events statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.EventsStats"
                        }
                    }
                }
            }
        },
        "/admin/events/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Возвращает полную информацию о мероприятии, включая статистику бронирований. / [EN] Returns full event information including booking stats.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-events"
                ],
                "summary": "[RU] Детальная информация о мероприятии / [EN] Event details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID мероприятия / [EN] Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.EventDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Принудительное обновление данных мероприятия администратором. / [EN] Forced update of event data by admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-events"
                ],
                "summary": "[RU] Обновить мероприятие (админ) / [EN] Update event (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID мероприятия / [EN] Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "[RU] Данные для обновления / [EN] Update data",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.UpdateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Event"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Переводит мероприятие в статус CANCELLED. / [EN] Changes event status to CANCELLED.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-events"
                ],
                "summary": "[RU] Отменить/Удалить мероприятие / [EN] Cancel/Delete event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID мероприятия / [EN] Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/events/{id}/publish": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Опубликовать мероприятие от лица администратора. / [EN] Publish event as admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-events"
                ],
                "summary": "[RU] Опубликовать мероприятие (админ) / [EN] Publish event (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID мероприятия / [EN] Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Event"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.Event": {
            "type": "object",
            "properties": {
                "available_seats": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "event_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "organizer_email": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.EventStatus"
                },
                "title": {
                    "type": "string"
                },
                "total_seats": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
        "github_com_bekzat-kamen_booking_system_api_internal_model.EventDetail": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Event"
                },
                "statistics": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "github_com_bekzat-kamen_booking_system_api_internal_model.EventStatus": {
            "type": "string",
            "enum": [
                "draft",
                "published",
                "sold_out",
                "cancelled",
                "completed"
            ],
            "x-enum-varnames": [
                "EventStatusDraft",
                "EventStatusPublished",
                "EventStatusSoldOut",
                "EventStatusCancelled",
                "EventStatusCompleted"
            ]
        },
        "github_com_bekzat-kamen_booking_system_api_internal_model.EventsStats": {
            "type": "object",
            "properties": {
                "cancelled_events": {
                    "type": "integer"
                },
                "draft_events": {
                    "type": "integer"
                },
                "published_events": {
                    "type": "integer"
                },
                "total_events": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.UpdateEventRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 5000
                },
                "event_date": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "sold_out",
                        "cancelled",
                        "completed"
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "total_seats": {
                    "type": "integer",
                    "minimum": 1
                },
                "venue": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}